@@ -4,10 +4,22 @@ import model.Match;
 import model.Player;
 
 import java.util.*;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
+/**
+ * Handles all pairing logic for a tournament, supporting both the Swiss System
+ * and Round Robin (Jeder-gegen-Jeden) modes.
+ *
+ * <p>In Swiss System mode, players are paired each round using backtracking to find
+ * valid pairings where no two players have already faced each other. Buchholz tiebreakers
+ * and points are used to order candidates. A bye (Freilos) is assigned when the player
+ * count is odd.</p>
+ *
+ * <p>In Round Robin mode, the circle method (Berger schedule) is used to generate one
+ * complete schedule for {@code n-1} rounds, guaranteeing every player meets every other
+ * player exactly once.</p>
+ */
 public class PairingEngine {
 
     private final List<Player> playerList;
@@ -15,45 +27,83 @@ public class PairingEngine {
     private final boolean modus;
     private final List<Match> matches;
     private final List<Match> allMatches;
-    private List<Player> byeList;
+    private Set<Player> byeList;
     private boolean finished;
 
+    /**
+     * Constructs a new PairingEngine.
+     *
+     * @param playerList  the list of players participating in the tournament
+     * @param tableNumber the number of available tables for match assignment
+     * @param modus       {@code true} for Round Robin, {@code false} for Swiss System
+     */
     public PairingEngine(List<Player> playerList, int tableNumber, boolean modus) {
         this.playerList = playerList;
         this.tableNumber = tableNumber;
         this.modus = modus;
         this.matches = new ArrayList<>();
         this.allMatches = new ArrayList<>();
-        this.byeList = new ArrayList<>();
+        this.byeList = new HashSet<>();
         this.finished = false;
     }
 
+    /**
+     * Returns the list of matches scheduled for the current round.
+     *
+     * @return the current round's match list
+     */
     public List<Match> getMatches() {
         return matches;
     }
 
+    /**
+     * Returns the cumulative list of all matches across every round played so far.
+     *
+     * @return the full match history including the current round
+     */
     public List<Match> getAllMatches() {
         return allMatches;
     }
 
+    /**
+     * Returns whether the tournament has been marked as finished.
+     *
+     * @return {@code true} if no further rounds can be generated
+     */
     public boolean isFinished() {
         return finished;
     }
 
+    /**
+     * Sets the finished state of the tournament.
+     *
+     * @param finished {@code true} to mark the tournament as finished
+     */
     public void setFinished(boolean finished) {
         this.finished = finished;
     }
 
+    /**
+     * Clears the match list for the current round, preparing for the next round's pairings.
+     */
     public void clearCurrentRound() {
         matches.clear();
     }
 
+    /**
+     * Restores the engine's state from previously saved match data.
+     * Replaces the in-memory match history and current-round matches with the provided lists,
+     * and rebuilds the bye list accordingly.
+     *
+     * @param savedAllMatches the full match history to restore
+     * @param savedMatches    the current round's matches to restore
+     */
     public void restoreState(List<Match> savedAllMatches, List<Match> savedMatches) {
         allMatches.clear();
         allMatches.addAll(savedAllMatches);
         matches.clear();
         matches.addAll(savedMatches);
-        byeList = new ArrayList<>(getPlayersWithBye(allMatches));
+        byeList = new HashSet<>(getPlayersWithBye(allMatches));
     }
 
     /**
@@ -65,37 +115,48 @@ public class PairingEngine {
      * @return pairing text to display, or {@code null} if no more rounds are possible
      */
     public String generatePairings(Set<Player> attemptedByePlayers, int currentRound) {
-        Set<Player> pairedPlayers = new HashSet<>();
         List<Integer> availableTables = IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList());
 
         if (modus) {
-            return generateParingsRoundRobin(pairedPlayers, availableTables, currentRound);
+            return generatePairingsRoundRobin(availableTables, currentRound);
         } else {
             return generatePairingsSwissSystem(attemptedByePlayers, new ArrayList<>(playerList), availableTables);
         }
     }
 
+    /**
+     * Generates Swiss System pairings for the current round.
+     * Assigns a bye if the player count is odd, then uses backtracking to find valid pairings.
+     * Falls back to a force-pairing strategy if backtracking yields no valid result.
+     *
+     * @param attemptedByePlayers players already tried as the bye recipient this round
+     * @param sortedList          the player list to pair, sorted by ranking
+     * @param availableTables     the pool of table numbers to assign
+     * @return formatted pairing text, or {@code null} if no valid pairings exist
+     */
     private String generatePairingsSwissSystem(Set<Player> attemptedByePlayers, List<Player> sortedList,
                                                List<Integer> availableTables) {
-        sortPlayers(sortedList);
+        sortPlayersByRanking(sortedList, modus);
         Player byePlayer = null;
-        String byeText = "";
 
         if (playerList.size() % 2 != 0) {
             byePlayer = assignByePlayer(sortedList, attemptedByePlayers);
             if (byePlayer == null) {
                 return forcePairing(availableTables);
             }
-            byeText = byePlayer.getFullName() + " (" + byePlayer.getClub() + ") - Freilos\n";
         }
+
+        Set<String> playedPairs = allMatches.stream()
+                .filter(m -> m.getSecondPlayer() != null)
+                .map(m -> createCanonicalPairing(m.getFirstPlayer(), m.getSecondPlayer()))
+                .collect(Collectors.toSet());
 
         List<Player> orderedPlayers = sortedList.stream()
                 .sorted(Comparator.comparingInt(Player::getPoints).reversed())
-                .collect(Collectors.toList());
+                .toList();
 
         List<Match> swissPairings = new ArrayList<>();
-        if (backtrackSwissPairing(orderedPlayers, new HashSet<>(), swissPairings)) {
-            StringBuilder pairingsText = new StringBuilder();
+        if (backtrackSwissPairing(orderedPlayers, new HashSet<>(), swissPairings, playedPairs)) {
             for (Match match : swissPairings) {
                 if (availableTables.isEmpty()) {
                     availableTables = IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList());
@@ -103,12 +164,8 @@ public class PairingEngine {
                 match.setTable(availableTables.removeFirst());
                 matches.add(match);
                 allMatches.add(match);
-                pairingsText.append(match.getFirstPlayer().getFullName()).append(" (").append(match.getFirstPlayer().getClub()).append(")")
-                        .append(" vs. ")
-                        .append(match.getSecondPlayer().getFullName()).append(" (").append(match.getSecondPlayer().getClub()).append(")")
-                        .append(" - Tisch ").append(match.getTableNumber()).append("\n");
             }
-            return pairingsText + byeText;
+            return formatMatchesAsText(matches);
         }
 
         if (byePlayer != null) {
@@ -118,23 +175,32 @@ public class PairingEngine {
         return forcePairing(availableTables);
     }
 
-    private boolean backtrackSwissPairing(List<Player> players, Set<Player> paired, List<Match> result) {
-        Player current = null;
-        for (Player p : players) {
-            if (!paired.contains(p)) {
-                current = p;
-                break;
-            }
-        }
+    /**
+     * Recursively pairs all remaining unpaired players using backtracking.
+     * At each step the first unpaired player is selected and matched against the best available
+     * opponent (ordered by closeness in points). Only pairings that still allow all remaining
+     * players to be fully matched are explored.
+     *
+     * @param players the complete ordered player list for this round
+     * @param paired  the set of players already assigned a match in the current recursion branch
+     * @param result  the accumulating list of confirmed match pairings
+     * @return {@code true} if a complete valid pairing was found, {@code false} otherwise
+     */
+    private boolean backtrackSwissPairing(List<Player> players, Set<Player> paired, List<Match> result,
+                                           Set<String> playedPairs) {
+        Player current = players.stream()
+                .filter(p -> !paired.contains(p))
+                .findFirst()
+                .orElse(null);
         if (current == null) return true;
 
-        for (Player opponent : getSwissOrderedOpponents(current, players, paired)) {
+        for (Player opponent : getSwissOrderedOpponents(current, players, paired, playedPairs)) {
             paired.add(current);
             paired.add(opponent);
 
-            if (canBeFullyMatched(players, paired)) {
+            if (canBeFullyMatched(players, paired, playedPairs)) {
                 result.add(new Match(current, opponent, -1));
-                if (backtrackSwissPairing(players, paired, result)) {
+                if (backtrackSwissPairing(players, paired, result, playedPairs)) {
                     return true;
                 }
                 result.removeLast();
@@ -146,92 +212,128 @@ public class PairingEngine {
         return false;
     }
 
-    private List<Player> getSwissOrderedOpponents(Player player, List<Player> players, Set<Player> paired) {
+    /**
+     * Returns a list of valid opponents for {@code player} in the Swiss System, ordered by
+     * closeness in tournament points. A player qualifies as a valid opponent if they have not
+     * yet been paired this round and have never played {@code player} before.
+     *
+     * @param player  the player to find opponents for
+     * @param players the full player list for this round
+     * @param paired  the set of players already paired in the current backtracking branch
+     * @return ordered list of eligible opponents
+     */
+    private List<Player> getSwissOrderedOpponents(Player player, List<Player> players, Set<Player> paired,
+                                                   Set<String> playedPairs) {
         return players.stream()
-                .filter(p -> !paired.contains(p) && !p.equals(player) && neverPlayedBefore(player, p))
+                .filter(p -> !paired.contains(p) && !p.equals(player)
+                        && !playedPairs.contains(createCanonicalPairing(player, p)))
                 .sorted(Comparator.comparingInt((Player p) -> Math.abs(p.getPoints() - player.getPoints())))
-                .collect(Collectors.toList());
+                .toList();
     }
 
-    private boolean canBeFullyMatched(List<Player> players, Set<Player> paired) {
+    /**
+     * Checks whether all currently unpaired players can still be matched given the set of already
+     * paired players. A player is considered matchable if at least one other unpaired player exists
+     * that they have not yet faced.
+     *
+     * @param players the full player list for this round
+     * @param paired  the set of players already assigned to matches in the current branch
+     * @return {@code true} if every remaining player has at least one valid opponent
+     */
+    private boolean canBeFullyMatched(List<Player> players, Set<Player> paired, Set<String> playedPairs) {
         List<Player> remaining = players.stream()
                 .filter(p -> !paired.contains(p))
                 .toList();
         for (Player player : remaining) {
             boolean hasOpponent = remaining.stream()
-                    .anyMatch(p -> !p.equals(player) && neverPlayedBefore(player, p));
+                    .anyMatch(p -> !p.equals(player) && !playedPairs.contains(createCanonicalPairing(player, p)));
             if (!hasOpponent) return false;
         }
         return true;
     }
 
-    private String generateParingsRoundRobin(Set<Player> pairedPlayers, List<Integer> availableTables, int currentRound) {
-        if (currentRound > playerList.size() - 1 + (playerList.size() % 2)) {
+    /**
+     * Generates Round Robin pairings using the circle method (Berger tables).
+     * <p>
+     * For n players, one player is fixed and the rest rotate by one position each round.
+     * For odd n, a virtual "bye" slot is added to make the count even; the player
+     * paired with the bye slot receives a Freilos in that round.
+     * </p>
+     */
+    private String generatePairingsRoundRobin(List<Integer> availableTables, int currentRound) {
+        int n = playerList.size();
+        int totalRounds = n - 1 + (n % 2);
+
+        if (currentRound > totalRounds) {
             finished = true;
             return null;
         }
 
-        Player byePlayer = null;
         List<Player> players = new ArrayList<>(playerList);
         players.sort(Comparator.comparing(Player::getTtr));
 
-        if (players.size() % 2 != 0) {
-            byePlayer = players.get(byeList.size());
-            byeList.add(byePlayer);
+        boolean oddPlayers = (n % 2 != 0);
+        List<Player> circle = new ArrayList<>(players);
+        if (oddPlayers) {
+            circle.add(null);
         }
 
-        if (byePlayer != null) {
-            players.remove(byePlayer);
+        int circleSize = circle.size();
+        Player fixedPlayer = circle.get(circleSize - 1);
+        List<Player> rotating = new ArrayList<>(circle.subList(0, circleSize - 1));
+
+        int r = currentRound - 1;
+        int rotLen = circleSize - 1;
+
+        List<Player[]> roundPairings = new ArrayList<>();
+        Player top = rotating.get(r % rotLen);
+        roundPairings.add(new Player[]{top, fixedPlayer});
+
+        for (int k = 1; k < circleSize / 2; k++) {
+            Player left = rotating.get((r + k) % rotLen);
+            Player right = rotating.get(((r - k) % rotLen + rotLen) % rotLen);
+            roundPairings.add(new Player[]{left, right});
         }
 
         StringBuilder pairingsText = new StringBuilder();
+        for (Player[] pair : roundPairings) {
+            Player p1 = pair[0];
+            Player p2 = pair[1];
 
-        for (int i = 0; i < players.size(); i++) {
-            Player player1 = players.get(i);
-            Player player2 = null;
-
-            for (int j = i + 1; j < players.size(); j++) {
-                Player opponent = players.get(j);
-                if (!neverPlayedBefore(player1, opponent)) {
-                    continue;
+            if (p1 == null || p2 == null) {
+                Player byePlayer = (p1 == null) ? p2 : p1;
+                Match byeMatch = new Match(byePlayer, null, -1);
+                matches.add(byeMatch);
+                allMatches.add(byeMatch);
+                pairingsText.append(byePlayer.getFullName()).append(" (").append(byePlayer.getClub()).append(")")
+                        .append(" - Freilos").append("\n");
+            } else {
+                if (availableTables.isEmpty()) {
+                    availableTables = IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList());
                 }
-                if (!pairedPlayers.contains(player1) && !pairedPlayers.contains(opponent)) {
-                    player2 = opponent;
-                }
+                int table = availableTables.removeFirst();
+                Match match = new Match(p1, p2, table);
+                matches.add(match);
+                allMatches.add(match);
+                pairingsText.append(p1.getFullName()).append(" (").append(p1.getClub()).append(")")
+                        .append(" vs. ")
+                        .append(p2.getFullName()).append(" (").append(p2.getClub()).append(")")
+                        .append(" - Tisch ").append(table).append("\n");
             }
-
-            if (player2 == null) {
-                continue;
-            }
-
-            if (availableTables.isEmpty()) {
-                availableTables = IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList());
-            }
-
-            int table = availableTables.removeFirst();
-            Match match = new Match(player1, player2, table);
-            matches.add(match);
-            allMatches.add(match);
-
-            pairingsText.append(player1.getFullName()).append(" (").append(player1.getClub()).append(")")
-                    .append(" vs. ")
-                    .append(player2.getFullName()).append(" (").append(player2.getClub()).append(")")
-                    .append(" - Tisch ").append(table).append("\n");
-
-            pairedPlayers.add(player1);
-            pairedPlayers.add(player2);
-        }
-
-        if (byePlayer != null) {
-            matches.add(new Match(byePlayer, null, -1));
-            allMatches.add(new Match(byePlayer, null, -1));
-            pairingsText.append(byePlayer.getFullName()).append(" (").append(byePlayer.getClub()).append(")")
-                    .append(" - Freilos").append("\n");
         }
 
         return pairingsText.toString();
     }
 
+    /**
+     * Fallback pairing strategy used when the standard Swiss backtracking finds no valid solution.
+     * Computes the set of all remaining unplayed pairings and selects a consistent assignment
+     * using a separate backtracking pass. Marks the tournament as finished if no valid assignment
+     * can be found.
+     *
+     * @param availableTables the pool of table numbers to assign to matches
+     * @return formatted pairing text, or {@code null} if the tournament is finished
+     */
     private String forcePairing(List<Integer> availableTables) {
         List<Match> openMatches = calculatePairingDifference(
                 generateAllPairings(playerList), allMatches, getPlayersWithBye(allMatches));
@@ -246,60 +348,58 @@ public class PairingEngine {
         matches.addAll(forcedMatches);
         allMatches.addAll(forcedMatches);
 
-        StringBuilder pairingsText = new StringBuilder();
-        for (Match match : matches) {
-            if (match.getSecondPlayer() == null) {
-                pairingsText.append(match.getFirstPlayer().getFullName())
-                        .append(" (").append(match.getFirstPlayer().getClub()).append(")")
-                        .append(" - Freilos\n");
-            } else {
-                pairingsText.append(match.getFirstPlayer().getFullName())
-                        .append(" (").append(match.getFirstPlayer().getClub()).append(")")
-                        .append(" vs. ")
-                        .append(match.getSecondPlayer().getFullName())
-                        .append(" (").append(match.getSecondPlayer().getClub()).append(")")
-                        .append(" - Tisch ").append(match.getTableNumber()).append("\n");
-            }
-        }
-        return pairingsText.toString();
+        return formatMatchesAsText(matches);
     }
 
+    /**
+     * Selects and assigns a bye (Freilos) player for this round.
+     * Picks the lowest-ranked player who has not yet had a bye and has not already been attempted
+     * as a bye candidate in the current round. The selected player is removed from
+     * {@code sortedList} and a bye match is added to both the current-round and full match lists.
+     *
+     * @param sortedList          the ranked player list; the bye player is removed in-place
+     * @param attemptedByePlayers players already tried as the bye recipient this round
+     * @return the chosen bye player, or {@code null} if none is eligible
+     */
     private Player assignByePlayer(List<Player> sortedList, Set<Player> attemptedByePlayers) {
-        for (Player player : sortedList) {
-            if (!byeList.contains(player) && !attemptedByePlayers.contains(player)) {
-                Match byeMatch = new Match(player, null, -1);
-                matches.add(byeMatch);
-                allMatches.add(byeMatch);
-                byeList = new ArrayList<>(getPlayersWithBye(allMatches));
-                sortedList.remove(player);
-                return player;
-            }
+        Player byePlayer = sortedList.stream()
+                .filter(p -> !byeList.contains(p) && !attemptedByePlayers.contains(p))
+                .findFirst()
+                .orElse(null);
+        if (byePlayer == null) {
+            return null;
         }
-        return null;
+        Match byeMatch = new Match(byePlayer, null, -1);
+        matches.add(byeMatch);
+        allMatches.add(byeMatch);
+        byeList = new HashSet<>(getPlayersWithBye(allMatches));
+        sortedList.remove(byePlayer);
+        return byePlayer;
     }
 
+    /**
+     * Removes the most recently added bye match for the given player from both the current-round
+     * and full match lists. Used when backtracking requires a different bye assignment.
+     *
+     * @param byePlayer the player whose bye match should be removed
+     */
     private void removeLastByeMatch(Player byePlayer) {
-        Match lastByeMatch = null;
-        for (Match match : matches) {
-            if (match.getFirstPlayer().equals(byePlayer) && match.getSecondPlayer() == null) {
-                lastByeMatch = match;
-                break;
-            }
-        }
-        if (lastByeMatch != null) {
-            matches.remove(lastByeMatch);
-            allMatches.remove(lastByeMatch);
-        }
+        matches.stream()
+                .filter(m -> m.getFirstPlayer().equals(byePlayer) && m.getSecondPlayer() == null)
+                .findFirst()
+                .ifPresent(m -> {
+                    matches.remove(m);
+                    allMatches.remove(m);
+                });
     }
 
-    private boolean neverPlayedBefore(Player player1, Player player2) {
-        return allMatches.stream()
-                .noneMatch(match ->
-                        (match.getFirstPlayer().equals(player1) && match.getSecondPlayer() != null && match.getSecondPlayer().equals(player2))
-                                || (match.getFirstPlayer().equals(player2) && match.getSecondPlayer() != null && match.getSecondPlayer().equals(player1)));
-    }
-
-    private void sortPlayers(List<Player> players) {
+    /**
+     * Sorts players by their ranking criteria. Used for both standings display and pairing order.
+     *
+     * @param players the list to sort (in place)
+     * @param modus   {@code true} for Round Robin, {@code false} for Swiss System
+     */
+    static void sortPlayersByRanking(List<Player> players, boolean modus) {
         if (!modus) {
             players.sort(Comparator.comparing(Player::getPoints)
                     .thenComparing(Player::getBuchholz)
@@ -315,6 +415,12 @@ public class PairingEngine {
         }
     }
 
+    /**
+     * Calculates all matches that could still be played in upcoming rounds, excluding matches
+     * already played (except those in the current round, which are still in progress).
+     *
+     * @return list of all unplayed possible pairings
+     */
     public List<Match> calculateAllPossibleOpenMatches() {
         List<Match> playedMatches = new ArrayList<>(allMatches);
         playedMatches.removeAll(matches);
@@ -338,35 +444,31 @@ public class PairingEngine {
             }
         }
 
-        AtomicReference<List<Integer>> availableTablesRef = new AtomicReference<>(
-                IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList()));
+        List<Integer> availableTablesRef = new ArrayList<>(
+                IntStream.rangeClosed(1, tableNumber).boxed().toList());
 
-        normalMatches.forEach(match -> {
-            if (availableTablesRef.get().isEmpty()) {
-                availableTablesRef.set(IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList()));
+        for (Match match : normalMatches) {
+            if (availableTablesRef.isEmpty()) {
+                availableTablesRef = new ArrayList<>(IntStream.rangeClosed(1, tableNumber).boxed().toList());
             }
-            match.setTable(availableTablesRef.get().removeFirst());
-        });
+            match.setTable(availableTablesRef.removeFirst());
+        }
 
         matches.addAll(normalMatches);
         allMatches.addAll(normalMatches);
         matches.addAll(byeMatches);
         allMatches.addAll(byeMatches);
 
-        StringBuilder text = new StringBuilder();
-        normalMatches.forEach(match ->
-                text.append(match.getFirstPlayer().getFullName()).append(" (").append(match.getFirstPlayer().getClub()).append(")")
-                        .append(" vs. ")
-                        .append(match.getSecondPlayer().getFullName()).append(" (").append(match.getSecondPlayer().getClub()).append(")")
-                        .append(" - Tisch ").append(match.getTableNumber()).append("\n")
-        );
-        byeMatches.forEach(match ->
-                text.append(match.getFirstPlayer().getFullName()).append(" - Freilos\n")
-        );
-
-        return text.toString();
+        return formatMatchesAsText(matches);
     }
 
+    /**
+     * Generates every possible unique pairing of players in the given list.
+     * If the number of players is odd, one bye match per player is also included.
+     *
+     * @param players the list of players to generate pairings for
+     * @return list of all possible matches (without duplicates)
+     */
     public List<Match> generateAllPairings(List<Player> players) {
         Set<String> uniquePairings = new HashSet<>();
         List<Match> allPairings = new ArrayList<>();
@@ -393,6 +495,15 @@ public class PairingEngine {
         return allPairings;
     }
 
+    /**
+     * Returns the subset of {@code allPairings} that have not yet been played and do not represent
+     * a bye for a player who has already received one.
+     *
+     * @param allPairings    the complete set of possible pairings
+     * @param playedPairings the matches that have already been played
+     * @param byePlayers     players who have already received a bye and should not receive another
+     * @return the remaining, unplayed pairings
+     */
     public List<Match> calculatePairingDifference(List<Match> allPairings, List<Match> playedPairings, List<Player> byePlayers) {
         Set<String> playedPairingsSet = playedPairings.stream()
                 .map(match -> createCanonicalPairing(match.getFirstPlayer(), match.getSecondPlayer()))
@@ -409,22 +520,38 @@ public class PairingEngine {
         return remainingPairings;
     }
 
+    /**
+     * Returns {@code true} if the match is a bye match and the player has already received a bye.
+     *
+     * @param match      the match to check
+     * @param byePlayers the list of players who have already had a bye
+     * @return {@code true} if the match is a duplicate bye for an already-exempted player
+     */
     private boolean isByeMatchForAnyPlayer(Match match, List<Player> byePlayers) {
         return match.getSecondPlayer() == null && byePlayers.contains(match.getFirstPlayer());
     }
 
+    /**
+     * Selects a set of matches from {@code matchList} such that every player appears in exactly
+     * one match. Table numbers are assigned to normal matches from {@code availableTables}.
+     * Returns {@code null} if no valid complete assignment covering all players is found.
+     *
+     * @param matchList       the pool of candidate matches to select from
+     * @param availableTables the pool of table numbers to assign
+     * @return a valid match assignment covering all players, or {@code null} if impossible
+     */
     public List<Match> selectUniquePlayerMatches(List<Match> matchList, List<Integer> availableTables) {
         List<Match> selectedMatches = new ArrayList<>();
         Set<Player> usedPlayers = new HashSet<>();
-        AtomicReference<List<Integer>> availableTablesAtomic = new AtomicReference<>(availableTables);
 
         if (backtrack(matchList, selectedMatches, usedPlayers, 0, false)) {
-            selectedMatches.forEach(match -> {
-                if (availableTablesAtomic.get().isEmpty()) {
-                    availableTablesAtomic.set(IntStream.rangeClosed(1, tableNumber).boxed().collect(Collectors.toList()));
+            List<Integer> tables = new ArrayList<>(availableTables);
+            for (Match match : selectedMatches) {
+                if (tables.isEmpty()) {
+                    tables = new ArrayList<>(IntStream.rangeClosed(1, tableNumber).boxed().toList());
                 }
-                match.setTable(availableTablesAtomic.get().removeFirst());
-            });
+                match.setTable(tables.removeFirst());
+            }
 
             if (!usedPlayers.equals(new HashSet<>(playerList))) {
                 return null;
@@ -436,6 +563,17 @@ public class PairingEngine {
         return null;
     }
 
+    /**
+     * Backtracking helper for {@link #selectUniquePlayerMatches}. Tries to build a valid
+     * complete match assignment from {@code matchList} starting at index {@code start}.
+     *
+     * @param matchList        the pool of candidate matches
+     * @param selectedMatches  accumulating list of selected matches
+     * @param usedPlayers      set of players already assigned to a match
+     * @param start            the current index in {@code matchList} to try
+     * @param byeMatchSelected whether a bye match has already been selected in this branch
+     * @return {@code true} if a complete valid assignment was found
+     */
     private boolean backtrack(List<Match> matchList, List<Match> selectedMatches, Set<Player> usedPlayers,
                                int start, boolean byeMatchSelected) {
         if (usedPlayers.size() == playerList.size()) {
@@ -483,24 +621,43 @@ public class PairingEngine {
         return false;
     }
 
+    /**
+     * Creates a collision-resistant canonical key for a player pair, independent of order.
+     * Uses null-character delimiters to prevent name-based ambiguity.
+     */
     private String createCanonicalPairing(Player player1, Player player2) {
-        String player1Name = player1.getFullName();
-        String player2Name = player2 != null ? player2.getFullName() : "Bye";
-
-        if (player1Name.compareTo(player2Name) < 0) {
-            return player1Name + "-" + player2Name;
-        } else {
-            return player2Name + "-" + player1Name;
-        }
+        String id1 = player1.getFullName() + "\0" + player1.getClub();
+        String id2 = player2 != null
+                ? player2.getFullName() + "\0" + player2.getClub()
+                : "\0BYE\0";
+        return id1.compareTo(id2) <= 0 ? id1 + "\1" + id2 : id2 + "\1" + id1;
     }
 
+    /**
+     * Returns a list of all players who have received a bye (Freilos) in the given match list.
+     *
+     * @param matchList the match list to inspect
+     * @return list of players with a bye; may be empty
+     */
     public List<Player> getPlayersWithBye(List<Match> matchList) {
-        Set<Player> playersWithBye = new HashSet<>();
-        for (Match match : matchList) {
-            if (match.getSecondPlayer() == null) {
-                playersWithBye.add(match.getFirstPlayer());
-            }
-        }
-        return new ArrayList<>(playersWithBye);
+        return matchList.stream()
+                .filter(m -> m.getSecondPlayer() == null)
+                .map(Match::getFirstPlayer)
+                .distinct()
+                .toList();
+    }
+
+    /**
+     * Formats a list of matches into display text for the pairings area.
+     */
+    public String formatMatchesAsText(List<Match> matchList) {
+        return matchList.stream()
+                .map(match -> match.getSecondPlayer() == null
+                        ? match.getFirstPlayer().getFullName() + " (" + match.getFirstPlayer().getClub() + ") - Freilos\n"
+                        : match.getFirstPlayer().getFullName() + " (" + match.getFirstPlayer().getClub() + ")"
+                                + " vs. "
+                                + match.getSecondPlayer().getFullName() + " (" + match.getSecondPlayer().getClub() + ")"
+                                + " - Tisch " + match.getTableNumber() + "\n")
+                .collect(Collectors.joining());
     }
 }
