@@ -11,17 +11,40 @@ import java.awt.print.PrinterException;
 import java.awt.print.PrinterJob;
 import java.util.List;
 
+/**
+ * Controller for the referee sheet (Schiedsrichterzettel) feature.
+ *
+ * <p>Builds the UI via {@link RefereeSheetsView} and provides the data and print logic
+ * for generating one referee sheet per match. Each sheet displays the two player names,
+ * five set rows, and a total-result row.</p>
+ */
 public record RefereeSheetsController(List<Match> matches) {
 
+    /**
+     * Constructs the controller for the given list of matches and schedules UI initialization
+     * on the Event Dispatch Thread.
+     *
+     * @param matches the list of matches (must not contain bye matches, i.e. second player != null)
+     */
     public RefereeSheetsController(List<Match> matches) {
         this.matches = matches;
         SwingUtilities.invokeLater(this::initializeUI);
     }
 
+    /**
+     * Initializes and displays the {@link RefereeSheetsView}.
+     */
     private void initializeUI() {
         new RefereeSheetsView(this, matches);
     }
 
+    /**
+     * Creates a read-only {@link DefaultTableModel} for the given match's referee sheet.
+     * The model contains a header row with player names and rows for each set plus the total.
+     *
+     * @param match the match to create the sheet for
+     * @return the table model for the referee sheet
+     */
     public DefaultTableModel createRefereeSheetTableModel(Match match) {
         String[] columnNames = {"", match.getFirstPlayer().getFullName(), match.getSecondPlayer().getFullName()};
         Object[][] data = {
@@ -41,6 +64,10 @@ public record RefereeSheetsController(List<Match> matches) {
         };
     }
 
+    /**
+     * Opens the system print dialog and prints all referee sheets, fitting two sheets per page.
+     * Shows an error dialog if printing fails.
+     */
     public void printRefereeSheets() {
         PrinterJob job = PrinterJob.getPrinterJob();
         job.setJobName("Schiedsrichterzettel");
@@ -78,11 +105,23 @@ public record RefereeSheetsController(List<Match> matches) {
         if (job.printDialog()) {
             try {
                 job.print();
-            } catch (PrinterException ignored) {
+            } catch (PrinterException e) {
+                JOptionPane.showMessageDialog(null, "Fehler beim Drucken der Schiedsrichterzettel.", "Druckfehler", JOptionPane.ERROR_MESSAGE);
             }
         }
     }
 
+    /**
+     * Draws a single referee sheet for the given match at the specified vertical offset.
+     * The sheet includes a header with table number and player names, five set rows,
+     * and a highlighted total-result row.
+     *
+     * @param g2     the graphics context to draw into
+     * @param match  the match whose data is rendered
+     * @param width  the available printable width in points
+     * @param height the allocated height for this sheet in points
+     * @param y      the vertical offset (top edge) at which to draw the sheet
+     */
     private void drawRefereeSheet(Graphics2D g2, Match match, int width, int height, int y) {
         final int margin = 8;
         final int innerWidth = width - margin * 2;
@@ -170,6 +209,15 @@ public record RefereeSheetsController(List<Match> matches) {
         g2.drawRect(margin, y, innerWidth, height - 4);
     }
 
+    /**
+     * Truncates {@code text} to fit within {@code maxWidth} pixels as measured by {@code fm},
+     * appending an ellipsis ({@code ...}) if truncation is necessary.
+     *
+     * @param text     the text to potentially truncate
+     * @param fm       the font metrics used to measure string width
+     * @param maxWidth the maximum allowed width in pixels
+     * @return the original text if it fits, or a truncated version with {@code ...} appended
+     */
     private String truncate(String text, FontMetrics fm, int maxWidth) {
         if (fm.stringWidth(text) <= maxWidth) return text;
         String ellipsis = "...";
